requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0
anthropic>=0.40.0
openai>=1.0.0
aiohttp>=3.9.0
httpx>=0.27.0
reportlab>=4.0
aiosmtplib>=3.0
apscheduler>=3.10
